Better base class structure.
