+++ src/Task/Ecs/EcsTask.php
@@ -5,14 +5,14 @@ declare(strict_types = 1);
 namespace Wunderio\GrumPHP\Task\Ecs;
 
 use GrumPHP\Collection\ProcessArgumentsCollection;
-use Wunderio\GrumPHP\Task\AbstractConfigurableContextFileExternalTask;
+use Wunderio\GrumPHP\Task\AbstractMultiPathProcessingTask;
 
 /**
  * Class EcsTask.
  *
  * @package Wunderio\GrumPHP\Task\Ecs
  */
-class EcsTask extends AbstractConfigurableContextFileExternalTask {
+class EcsTask extends AbstractMultiPathProcessingTask {
 
   /**
    * {@inheritdoc}
