+++ src/Task/PhpCheckSyntax/PhpCheckSyntaxTask.php
@@ -5,24 +5,22 @@ declare(strict_types = 1);
 namespace Wunderio\GrumPHP\Task\PhpCheckSyntax;
 
 use GrumPHP\Collection\ProcessArgumentsCollection;
-use Wunderio\GrumPHP\Task\AbstractPerPathExternalTask;
+use Wunderio\GrumPHP\Task\AbstractSinglePathProcessingTask;
 
 /**
  * Class PhpCheckSyntaxTask.
  *
  * @package Wunderio\GrumPHP\Task\PhpCheckSyntaxTask
  */
-class PhpCheckSyntaxTask extends AbstractPerPathExternalTask {
+class PhpCheckSyntaxTask extends AbstractSinglePathProcessingTask {
 
   /**
    * {@inheritdoc}
    */
-  public function buildArguments(iterable $files): ProcessArgumentsCollection {
+  public function buildArgumentsFromPath(string $path): ProcessArgumentsCollection {
     $arguments = $this->processBuilder->createArgumentsForCommand('php');
     $arguments->add('-l');
-    foreach ($files as $file) {
-      $arguments->add($file);
-    }
+    $arguments->add($path);
 
     return $arguments;
   }
